requirements.txt
fastapi==0.75.0
psycopg2-binary==2.9.3
uvicorn==0.17.6
python-dotenv==0.19.2
pandas==1.4.2
catboost==1.0.6
sqlalchemy==1.4.39
